pyproject.toml
[project]
name = "face-recognition-classifier"
version = "0.1.0"
description = "Lightweight face recognition and classification without training from scratch."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "insightface>=0.7.3",
    "numpy>=1.24",
    "opencv-python>=4.8",
    "rich>=13.7",
    "tqdm>=4.66",
]

[project.optional-dependencies]
cpu = [
    "onnxruntime>=1.17",
]
gpu = [
    "onnxruntime-gpu>=1.17",
]
mqtt = [
    "paho-mqtt>=2.0",
]

[project.scripts]
face-id = "face_id.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/face_id"]
